requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
Django>=4.2.0
django-allauth==65.7.0
dnspython==2.7.0
exa-py>=0.1.0
exceptiongroup==1.2.2
Flask==3.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
lxml==5.4.0
MarkupSafe==3.0.2
mongoengine==0.29.1
openai>=1.0.0
packaging==25.0
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
pytest==8.3.5
pytest-mock==3.14.0
python-dotenv>=1.0.0
python-openid==2.2.5
python3-saml==1.16.0
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
xmlsec==1.3.15
zipp==3.21.0
gunicorn>=21.2.0
django-cors-headers>=4.0.0  # For PostgreSQL support
whitenoise>=6.0.0  # For serving static files
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
social-auth-app-django==5.4.0
python-jose==3.3.0
